update config tsconfig fix
