--- frontend/src/Types/ProjectsType.ts
@@ -1,42 +1,0 @@
-export interface ICreateProject {
-  name: string;
-  description: string;
-  email_client: string;
-  price_hour: number;
-}
-
-export interface IShowProject {
-  name: string,
-  email_client: string,
-  init_date: null,
-  end_date: null,
-  is_completed: boolean,
-  price_hour: string
-}
-
-export interface IClient {
-  avatar: null,
-  birthdate: string,
-  email: string,
-  first_name: string,
-  last_name: string,
-  phone: string,
-  uuid: string,
-  role_id: number
-}
-
-export interface IProjectResponse {
-  uuid?:            string;
-  init_date:       null;
-  end_date:        null;
-  name:            string;
-  price_hour:      string;
-  description?:     string;
-  id_cliente?:      null | string;
-  email_client:    string;
-  is_completed:    boolean;
-  is_active?:       boolean;
-  status_uuid?:     string;
-  custom_label_id?: null;
-  user_uuid?:       string;
-}
