--- frontend/src/Types/types.d.ts
@@ -1,13 +1,0 @@
-export interface StatusType {
-  uuid: string
-  color: string
-  name: string
-}
-
-export interface LabelType {
-  uuid: string
-  description: string
-  color: string
-  name: string
-  is_active: boolean
-}
