+++ frontend/src/components/Clients/Client.tsx
@@ -3,7 +3,7 @@ import { Box, Button } from '@mui/material';
 import PersonAddAlt1Icon from '@mui/icons-material/PersonAddAlt1';
 import axios from 'axios';
 import { useEffect, useState } from 'react';
-import { Person, Result } from '../../Types/person';
+import type { Person, Result } from 'types.d.ts';
 import { TableClient } from './TableClient';
 
 export const Client = () => {
