+++ frontend/src/components/Projects/components/TableProject.tsx
@@ -1,4 +1,4 @@
-import type { IProjectResponse } from "../../../types/ProjectsType";
+import type { IProjectResponse } from "types.d.ts";
