+++ frontend/tsconfig.json
@@ -1,5 +1,13 @@
 {
   "compilerOptions": {
+    "baseUrl": "src",  // Establece la base de las rutas como el directorio src
+      "paths": {
+        "@types/*": ["types/*"],  // Define un alias @types para la carpeta types
+        "@components/*": ["components/*"],  // Alias para la carpeta components
+        "@hooks/*": ["hooks/*"],  // Alias para la carpeta hooks
+        "@routes/*": ["routes/*"],  // Alias para la carpeta routes
+        "@utils/*": ["utils/*"]  // Alias para la carpeta utils
+       },
     "target": "ES2020",
     "useDefineForClassFields": true,
     "lib": ["ES2020", "DOM", "DOM.Iterable"],
